Move IrradianceVolume to bevy_light
