+++ crates/bevy_light/src/probe.rs
@@ -107,3 +107,50 @@ impl Default for EnvironmentMapLight {
         }
     }
 }
+
+/// The component that defines an irradiance volume.
+///
+/// See `bevy_pbr::irradiance_volume` for detailed information.
+///
+/// This component requires the [`LightProbe`] component, and is typically used with
+/// [`bevy_transform::components::Transform`] to place the volume appropriately.
+#[derive(Clone, Reflect, Component, Debug)]
+#[reflect(Component, Default, Debug, Clone)]
+#[require(LightProbe)]
+pub struct IrradianceVolume {
+    /// The 3D texture that represents the ambient cubes, encoded in the format
+    /// described in `bevy_pbr::irradiance_volume`.
+    pub voxels: Handle<Image>,
+
+    /// Scale factor applied to the diffuse and specular light generated by this component.
+    ///
+    /// After applying this multiplier, the resulting values should
+    /// be in units of [cd/m^2](https://en.wikipedia.org/wiki/Candela_per_square_metre).
+    ///
+    /// See also <https://google.github.io/filament/Filament.html#lighting/imagebasedlights/iblunit>.
+    pub intensity: f32,
+
+    /// Whether the light from this irradiance volume has an effect on meshes
+    /// with lightmaps.
+    ///
+    /// Set this to false if your lightmap baking tool bakes the light from this
+    /// irradiance volume into the lightmaps in order to avoid counting the
+    /// irradiance twice. Frequently, applications use irradiance volumes as a
+    /// lower-quality alternative to lightmaps for capturing indirect
+    /// illumination on dynamic objects, and such applications will want to set
+    /// this value to false.
+    ///
+    /// By default, this is set to true.
+    pub affects_lightmapped_meshes: bool,
+}
+
+impl Default for IrradianceVolume {
+    #[inline]
+    fn default() -> Self {
+        IrradianceVolume {
+            voxels: Handle::default(),
+            intensity: 0.0,
+            affects_lightmapped_meshes: true,
+        }
+    }
+}
