+++ crates/bevy_pbr/src/light_probe/irradiance_volume.rs
@@ -133,9 +133,8 @@
 //!
 //! [Why ambient cubes?]: #why-ambient-cubes
 
-use bevy_ecs::{component::Component, reflect::ReflectComponent};
 use bevy_image::Image;
-use bevy_light::LightProbe;
+pub use bevy_light::IrradianceVolume;
 use bevy_render::{
     render_asset::RenderAssets,
     render_resource::{
@@ -145,11 +144,9 @@ use bevy_render::{
     renderer::{RenderAdapter, RenderDevice},
     texture::{FallbackImage, GpuImage},
 };
-use bevy_utils::default;
 use core::{num::NonZero, ops::Deref};
 
-use bevy_asset::{AssetId, Handle};
-use bevy_reflect::{std_traits::ReflectDefault, Reflect};
+use bevy_asset::AssetId;
 
 use crate::{
     add_cubemap_texture_view, binding_arrays_are_usable, RenderViewLightProbes,
@@ -162,53 +159,6 @@ use super::LightProbeComponent;
 /// overflow the number of texture bindings when deferred rendering is in use
 /// (see issue #11885).
 pub(crate) const IRRADIANCE_VOLUMES_ARE_USABLE: bool = cfg!(not(target_arch = "wasm32"));
-
-/// The component that defines an irradiance volume.
-///
-/// See [`crate::irradiance_volume`] for detailed information.
-///
-/// This component requires the [`LightProbe`] component, and is typically used with
-/// [`bevy_transform::components::Transform`] to place the volume appropriately.
-#[derive(Clone, Reflect, Component, Debug)]
-#[reflect(Component, Default, Debug, Clone)]
-#[require(LightProbe)]
-pub struct IrradianceVolume {
-    /// The 3D texture that represents the ambient cubes, encoded in the format
-    /// described in [`crate::irradiance_volume`].
-    pub voxels: Handle<Image>,
-
-    /// Scale factor applied to the diffuse and specular light generated by this component.
-    ///
-    /// After applying this multiplier, the resulting values should
-    /// be in units of [cd/m^2](https://en.wikipedia.org/wiki/Candela_per_square_metre).
-    ///
-    /// See also <https://google.github.io/filament/Filament.html#lighting/imagebasedlights/iblunit>.
-    pub intensity: f32,
-
-    /// Whether the light from this irradiance volume has an effect on meshes
-    /// with lightmaps.
-    ///
-    /// Set this to false if your lightmap baking tool bakes the light from this
-    /// irradiance volume into the lightmaps in order to avoid counting the
-    /// irradiance twice. Frequently, applications use irradiance volumes as a
-    /// lower-quality alternative to lightmaps for capturing indirect
-    /// illumination on dynamic objects, and such applications will want to set
-    /// this value to false.
-    ///
-    /// By default, this is set to true.
-    pub affects_lightmapped_meshes: bool,
-}
-
-impl Default for IrradianceVolume {
-    #[inline]
-    fn default() -> Self {
-        IrradianceVolume {
-            voxels: default(),
-            intensity: 0.0,
-            affects_lightmapped_meshes: true,
-        }
-    }
-}
 
 /// All the bind group entries necessary for PBR shaders to access the
 /// irradiance volumes exposed to a view.
