+++ crates/bevy_pbr/src/light_probe/mod.rs
@@ -288,8 +288,7 @@ impl Plugin for LightProbePlugin {
         load_shader_library!(app, "environment_map.wgsl");
         load_shader_library!(app, "irradiance_volume.wgsl");
 
-        app.register_type::<IrradianceVolume>()
-            .add_plugins(ExtractInstancesPlugin::<EnvironmentMapIds>::new());
+        app.add_plugins(ExtractInstancesPlugin::<EnvironmentMapIds>::new());
     }
 
     fn finish(&self, app: &mut App) {
